requirements.txt
Flask==2.2.5
pytesseract==0.3.10
Pillow==9.5.0
pdfplumber==0.10.0
python-docx==0.8.11
